Tutorials/client/requirements.txt
awscrt==0.20.10
awsiotsdk==1.21.5

